+++ CCP/plugin/network_service/fermat-ccp-plugin-network-service-crypto-transmission2-bitdubai/src/main/java/bitdubai/version_1/database/communications/CryptoTransmissionMetadataDAO_V2.java
@@ -9,7 +9,6 @@ import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseTable;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseTableFilter;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseTableRecord;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseTransaction;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.PluginDatabaseSystem;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantDeleteRecordException;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantInsertRecordException;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantLoadTableToMemoryException;
@@ -31,9 +30,9 @@ import bitdubai.version_1.exceptions.CantDeleteRecordDataBaseException;
 import bitdubai.version_1.exceptions.CantGetCryptoTransmissionMetadataException;
 import bitdubai.version_1.exceptions.CantSaveCryptoTransmissionMetadatatException;
 import bitdubai.version_1.exceptions.CantUpdateRecordDataBaseException;
-import bitdubai.version_1.structure.structure.CryptoTransmissionMessageType;
-import bitdubai.version_1.structure.structure.CryptoTransmissionMetadataRecord;
-import bitdubai.version_1.structure.structure.CryptoTransmissionResponseMessage;
+import bitdubai.version_1.structure.CryptoTransmissionMessageType;
+import bitdubai.version_1.structure.CryptoTransmissionMetadataRecord;
+import bitdubai.version_1.structure.CryptoTransmissionResponseMessage;
 
 /**
  * Created by Matias Furszyfer on 2015.10.04..
