+++ CCP/plugin/network_service/fermat-ccp-plugin-network-service-crypto-transmission2-bitdubai/src/main/java/bitdubai/version_1/structure/CryptoTransmissionMessageType.java
@@ -1,4 +1,4 @@
-package bitdubai.version_1.structure.structure;
+package bitdubai.version_1.structure;
 
 import com.bitdubai.fermat_api.layer.all_definition.exceptions.InvalidParameterException;
