+++ CCP/plugin/network_service/fermat-ccp-plugin-network-service-crypto-transmission2-bitdubai/src/main/java/bitdubai/version_1/structure/CryptoTransmissionMetadataRecord.java
@@ -1,4 +1,4 @@
-package bitdubai.version_1.structure.structure;
+package bitdubai.version_1.structure;
 
 import com.bitdubai.fermat_api.layer.all_definition.enums.CryptoCurrency;
 import com.bitdubai.fermat_ccp_api.layer.network_service.crypto_transmission.enums.CryptoTransmissionMetadataState;
@@ -20,7 +20,7 @@ public class CryptoTransmissionMetadataRecord extends CryptoTransmissionMessage 
     private String paymentDescription;
     private CryptoTransmissionMetadataType cryptoTransmissionMetadataType;
     private long timestamp;
-    public CryptoTransmissionMetadataRecord(UUID transactionId, CryptoTransmissionMessageType cryptoTransmissionMessage, UUID requestId, CryptoCurrency cryptoCurrency, long cryptoAmount, String senderPublicKey, String destinationPublicKey, String associatedCryptoTransactionHash, String paymentDescription, CryptoTransmissionProtocolState cryptoTransmissionProtocolState, CryptoTransmissionMetadataType cryptoTransmissionMetadataType, long timestamp, boolean pendingFlag, int sentCount, CryptoTransmissionMetadataState cryptoTransmissionMetadataState) {
+    public CryptoTransmissionMetadataRecord(UUID transactionId, bitdubai.version_1.structure.CryptoTransmissionMessageType cryptoTransmissionMessage, UUID requestId, CryptoCurrency cryptoCurrency, long cryptoAmount, String senderPublicKey, String destinationPublicKey, String associatedCryptoTransactionHash, String paymentDescription, CryptoTransmissionProtocolState cryptoTransmissionProtocolState, CryptoTransmissionMetadataType cryptoTransmissionMetadataType, long timestamp, boolean pendingFlag, int sentCount, CryptoTransmissionMetadataState cryptoTransmissionMetadataState) {
         super(transactionId, cryptoTransmissionMessage,senderPublicKey,destinationPublicKey,cryptoTransmissionProtocolState,cryptoTransmissionMetadataState,pendingFlag,sentCount);
         this.requestId = requestId;
         this.cryptoCurrency = cryptoCurrency;
